feat(client): made mcp json editor into shared components
