+++ libs/shared/src/components/index.ts
@@ -2,3 +2,4 @@ export * from "./auditlog";
 export * from "./engine";
 export * from "./flex-layout";
 export * from "./form";
+export * from "./mcp-json-editor";
